+++ v2/tweet_options.go
@@ -42,65 +42,6 @@ func (t TweetLookupOpts) addQuery(req *http.Request) {
 	}
 }
 
-// TweetRecentSearchOpts are the optional parameters for the recent seach API
-type TweetRecentSearchOpts struct {
-	Expansions  []Expansion
-	MediaFields []MediaField
-	PlaceFields []PlaceField
-	PollFields  []PollField
-	TweetFields []TweetField
-	UserFields  []UserField
-	StartTime   time.Time
-	EndTime     time.Time
-	MaxResults  int
-	NextToken   string
-	SinceID     string
-	UntilID     string
-}
-
-func (t TweetRecentSearchOpts) addQuery(req *http.Request) {
-	q := req.URL.Query()
-	if len(t.Expansions) > 0 {
-		q.Add("expansions", strings.Join(expansionStringArray(t.Expansions), ","))
-	}
-	if len(t.MediaFields) > 0 {
-		q.Add("media.fields", strings.Join(mediaFieldStringArray(t.MediaFields), ","))
-	}
-	if len(t.PlaceFields) > 0 {
-		q.Add("place.fields", strings.Join(placeFieldStringArray(t.PlaceFields), ","))
-	}
-	if len(t.PollFields) > 0 {
-		q.Add("poll.fields", strings.Join(pollFieldStringArray(t.PollFields), ","))
-	}
-	if len(t.TweetFields) > 0 {
-		q.Add("tweet.fields", strings.Join(tweetFieldStringArray(t.TweetFields), ","))
-	}
-	if len(t.UserFields) > 0 {
-		q.Add("user.fields", strings.Join(userFieldStringArray(t.UserFields), ","))
-	}
-	if !t.StartTime.IsZero() {
-		q.Add("start_time", t.StartTime.Format(time.RFC3339))
-	}
-	if !t.EndTime.IsZero() {
-		q.Add("end_time", t.EndTime.Format(time.RFC3339))
-	}
-	if t.MaxResults > 0 {
-		q.Add("max_results", strconv.Itoa(t.MaxResults))
-	}
-	if len(t.NextToken) > 0 {
-		q.Add("next_token", t.NextToken)
-	}
-	if len(t.SinceID) > 0 {
-		q.Add("since_id", t.SinceID)
-	}
-	if len(t.UntilID) > 0 {
-		q.Add("until_id", t.UntilID)
-	}
-	if len(q) > 0 {
-		req.URL.RawQuery = q.Encode()
-	}
-}
-
 // UserTweetTimelineOpts are the options for the user tweet timeline request
 type UserTweetTimelineOpts struct {
 	Expansions      []Expansion
